Split codec package into transport independent and dependent elements.
Any transport will need the ability to convert instances of PaxosMessage into
byte[] that can be shunted over the wire and thus this element should be
generic. This will typically leave an implementer a trivial wrapper to write.

In cases where specific, tight control over transport marshalling is required,
the implementer will need to write their own codecs using the generic one's as
a template.