Initial protocol overhaul.
The previous protocol had a bunch of faults in particular the logic around use
of collects was broken. They must always be executed but in the majority of
cases ALs will return NO_VALUE allowing the leader to quickly complete a paxos
instance rather than perform recovery actions. This is most effective when the
leader is not constantly switching as implemented (correctly) in the old
version of the protocol.

This revision now correctly issues collects and processes the resulting LASTs.
First class support for handling consolidated handbacks and values was also
added to eliminate a source of bugs possible with naked array handling and to
make the code more readable with fewer exceptions. High watermarks have also
been eliminated as the revised protocol implementation can handle progressing
through a number of already completed instances before dispatching the client's
request that triggered the leader.

Next steps are:

(1) To have ALs signal a leader when it's asking about a paxos instance that
is too out of date and for the leader to react by signalling clients to move to
another leader and asking the AL to dispatch a recovery.
(2) AL must also be fixed to consult its log for an old value as part of the
out of date changes.
(3) Implement buffering of completion notifications from AL to cope with out
of order delivery where AL becomes aware of a paxos instance result that is
some number of instances since the last one seen. This mod will allow the AL
to process the instance, send a notification for it and dispatch recovery
requests to other nodes and process those results also.
(4) Have the buffering in (3) support clearing of all messages in the case of
full recovery from a new database and logs obtained from another node. This
work will also include adding an Event that handles delivering of new initial
state to a state machine.
(5) Checkpoint which will entail another kind of event to request the current
state from a state machine.