Fix leader to reuse its round number when appropriate.
Since the initial protocol rewrite, leader has been constantly incrementing its
round number which means additional disk writes we'd like to avoid.

In course of making this change, a number of other bugs were highlighted. The
leader was not correctly handling oldround messages continuing to retry and
ultimately exhausting memory instead of giving up with a redirect for the
client. Secondly, the acceptor/learner was not correctly handling a collect
with the same round number from an existing leader and finally the
implementation of Collect.isInitial() was broken.