Make watermark a first class type in preparation for adding recovery.
Update the commentary on recovery.

Add some code to ignore success reports for paxos instances we've already
settled. Note that we take part in all other steps of the duplicate instance
to ensure the value is consistent with what has gone previously. This is
necessary in cases where some other node has a low watermark < ours and a
new leader has taken control and is in the process of recovering the Paxos
state.