Make transport usage in AcceptorLearner symmetric with Leader.
The AcceptorLearner ultimately must drive recovery/catchup and thus will need
to emit packets autonomously unprompted by a leader request. The previous
implementation made that impossible as AcceptorLearners had no access to an
instance of Transport with its responses being returned inline with the
request to process a received message and marshalled/sent but some external
entity.

This revision introduces a Transport instance to AcceptorLearners and provides
the benefit of simplifying the Mina implementation as the stacks are more
streamlined and initialisation is considerably less inter-dependent.