Remove packet buffer limit.
Size of the packet buffer must vary with the recovery effort. In essence its
size is defined by the window of sequence numbers that need recovery and how
many other requests are streaming in meantime. Limitations on the size of the
buffer will be by side-effect of other limits - e.g. on log size and checkpoint
windows.