Make AcceptorLearner properly implement Last response.
There are several concerns here, completed instances may be re-run by new
leaders to recover other replicas and the log must be consulted for those.
Active instance information is sourced from the packet buffer, it will be in
the log but it's not complete.

As a side-effect the logging infrastructure needed extending to return the
offset of the returned record and the Howl implementation required a bugfix
to return the appropriate field as opposed to all data. Success was modified
to include all the information required for a last to handle cases where begins
may have been missed or where the instance was already written to log.