Add the basic recovery skeleton.
This is not active as yet: AL becomes state machine with a recovery state and a
normal state. In recovery state, normal messages are queued, and only recovery
messages are handled. In normal state, recovery messages aren't expected and
transition to recovery occurs if the sequence number jumps to low_watermark + 2
or more.