Add support for Netty.
The key benefit of using Netty is that it permits full use of multicast which
amongst other things allows for a full remote simulation inside of one JVM
which is not possible with MINA and its UDP/Broadcast approach where only one
client sees the packets.

This will make it possible to eliminate much of the local test infrastructure
and provide the majority of infrastructure required for the ultimate
paxos network service.