Implement first-stage recovery.
AL will now detect it's lagging and ask for a node (the leader) to catch it up.

Still to implement is checkpointing which dictates the minimum paxos instance
that can be recovered. When this happens, recovery will need modifying to
emit an abort should an AL ask for catchup below that waterline. AL will then
need to pull log and checkpoint files across and restart.